update StackView layout
